+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openparser/cli
 
+## 1.0.8
+
+### Patch Changes
+
+- @openparser/sdk@1.0.8
+
 ## 1.0.7
 
 ### Patch Changes